Crate bn254_blackbox_solver

Re-exports

pub use poseidon2_constants::POSEIDON2_CONFIG;

pub use poseidon2_constants::POSEIDON2_CONFIG;

Modules

embedded_curve_ops 🔒

generator 🔒

This module is adapted from the Barustenberg Rust implementation of the Barretenberg library.

poseidon2 🔒

poseidon2_constants 🔒

Structs

Bn254BlackBoxSolver

POSEIDON2_CONFIG

The complete Poseidon2 configuration for BN254. This includes all parameters and constants needed for the permutation.

Poseidon2Config

Configuration parameters for Poseidon2 permutation

Poseidon2Sponge

Functions

derive_generators

Derives generator points.

embedded_curve_add

field_from_hex

Helper function to convert hex strings to field elements

multi_scalar_mul

Performs multi scalar multiplication of points with scalars.

poseidon2_permutation

poseidon_hash

Performs a poseidon hash with a sponge construction equivalent to the one in the Barretenberg proving system

Type Aliases

FieldElement 🔒