bn254_blackbox_solver::generator::generators

Function derive_generators

pub fn derive_generators(
    domain_separator_bytes: &[u8],
    num_generators: u32,
    starting_index: u32,
) -> Vec<Affine> 

Derives generator points.

ALGORITHM DESCRIPTION

Each generator has an associated “generator index” described by its location in the vector

a 64-byte preimage buffer is generated with the following structure:

bytes 0-31: BLAKE3 hash of domain_separator

bytes 32-63: generator index in big-endian form

The hash-to-curve algorithm is used to hash the above into a curve point.

NOTE: The domain separator is included to ensure that it is possible to derive independent sets of index-addressable generators.