bn254_blackbox_solver::generator::hash_to_curve
Function hash_to_curve

pub(crate) fn hash_to_curve(seed: &[u8], attempt_count: u8) -> Affine

Hash a seed buffer into a point

ALGORITHM DESCRIPTION

Initialize unsigned integer attempt_count = 0

Copy seed into a buffer whose size is 2 bytes greater than seed (initialized to 0)

Interpret attempt_count as a byte and write into buffer at [buffer.size() - 2]

Compute Blake3 hash of buffer

Set the end byte of the buffer to 1

Compute Blake3 hash of buffer

Interpret the two hash outputs as the high / low 256 bits of a 512-bit integer (big-endian)

Derive x-coordinate of point by reducing the 512-bit integer modulo the curve’s field modulus (Fq)

Compute y^2 from the curve formula y^2 = x^3 + ax + b (a, b are curve params. for BN254, a = 0, b = 3)

IF y^2 IS NOT A QUADRATIC RESIDUE:

a. increment attempt_count by 1 and go to step 2

IF y^2 IS A QUADRATIC RESIDUE:

a. derive y coordinate via y = sqrt(y)

b. Interpret most significant bit of 512-bit integer as a ‘parity’ bit

c. If parity bit is set AND y’s most significant bit is not set, invert y

d. If parity bit is not set AND y’s most significant bit is set, invert y

e. return (x, y)

N.B. steps c. and e. are because the sqrt() algorithm can return 2 values, we need to a way to canonically distinguish between these 2 values and select a “preferred” one